# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
    title: Customer Service API
    description: Microservicio para la gestión de clientes bancarios
    version: 1.0.0
    contact:
        name: NTT Data Bootcamp
servers:
    - url: http://localhost:8080
      description: Servidor de desarrollo
...
paths:
    /enterprise:
        get:
            tags:
                - Enterprise Customers
            summary: Obtener todos los clientes empresariales
            description: Retorna una lista de todos los clientes empresariales
            operationId: retriveAllEnterpriseCustomers
            responses:
                '200':
                    description: Lista de clientes empresariales obtenida exitosamente
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/CustomerResponse'
                            example:
                                - id: "987fcdeb-51a2-43d1-b789-123456789abc"
                                  name: "Empresa ABC SAC"
                                  customerType: "ENTERPRISE"
                                  documentNumber: "20123456789"
                                  email: "contacto@empresaabc.com"
                                  active: true
                                  createdAt: "2024-01-15T10:00:00Z"
                                - id: "abc123de-f456-7890-abcd-ef1234567890"
                                  name: "Empresa XYZ SAC"
                                  customerType: "ENTERPRISE"
                                  documentNumber: "20987654321"
                                  email: "contacto@xyz.com"
                                  active: true
                                  createdAt: "2024-01-15T11:00:00Z"
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'

        post:
            tags:
                - Enterprise Customers
            summary: Crear cliente empresarial
            description: Crea un nuevo cliente empresarial
            operationId: createEnterpriseCustomer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateEnterpriseRequest'
                        example:
                            businessName: "Empresa ABC SAC"
                            ruc: "20123456789"
                            email: "contacto@empresaabc.com"
                            phone: "+51 1 1234567"
                            street: "Jr. Comercial 456"
                            city: "Lima"
                            country: "Perú"
                            zipCode: "15002"
            responses:
                '200':
                    description: Cliente empresarial creado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '400':
                    description: Datos de entrada inválidos
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /enterprise/{id}:
        get:
            tags:
                - Enterprise Customers
            summary: Obtener cliente empresarial por ID
            description: Obtiene un cliente empresarial específico por su ID único
            operationId: retrieveEnterpriseCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  description: ID único del cliente empresarial
                  schema:
                      type: string
                      format: uuid
                  example: "987fcdeb-51a2-43d1-b789-123456789abc"
            responses:
                '200':
                    description: Cliente empresarial obtenido exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente empresarial no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /enterprise/document-number/{documentNumber}:
        get:
            tags:
                - Enterprise Customers
            summary: Obtener cliente empresarial por documento
            description: Obtiene un cliente empresarial por tipo y número de documento
            operationId: retrieveEnterpriseCustomerByDocument
            parameters:
                - name: documentNumber
                  in: path
                  required: true
                  description: Número de documento
                  schema:
                      type: string
                      pattern: '^[0-9]{11}$'
                  example: "20123456789"
            responses:
                '200':
                    description: Cliente empresarial obtenido exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente empresarial no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /enterprise/update/{id}:
        put:
            tags:
                - Enterprise Customers
            summary: Actualizar cliente empresarial
            description: Actualiza la información de un cliente empresarial existente
            operationId: updateEnterpriseCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  description: ID único del cliente empresarial
                  schema:
                      type: string
                      format: uuid
                  example: "987fcdeb-51a2-43d1-b789-123456789abc"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateEnterpriseRequest'
                        example:
                            businessName: "Empresa ABC SAC Actualizada"
                            email: "nuevo@empresaabc.com"
                            phone: "+51 1 9876543"
                            street: "Jr. Nueva Dirección 789"
                            city: "Trujillo"
                            country: "Perú"
                            zipCode: "13001"
            responses:
                '200':
                    description: Cliente empresarial actualizado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '400':
                    description: Datos de entrada inválidos
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: Cliente empresarial no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /enterprise/desactivate/{id}:
        delete:
            tags:
                - Enterprise Customers
            summary: Desactivar cliente empresarial
            description: Desactiva un cliente empresarial (soft delete)
            operationId: desactivateEnterpriseCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  description: ID único del cliente empresarial
                  schema:
                      type: string
                      format: uuid
                  example: "987fcdeb-51a2-43d1-b789-123456789abc"
            responses:
                '200':
                    description: Cliente empresarial desactivado exitosamente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CustomerResponse'
                '404':
                    description: Cliente empresarial no encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '500':
                    description: Error interno del servidor
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
...
components:
    schemas:
...
        CreateEnterpriseRequest:
            type: object
            required:
                - businessName
                - ruc
                - email
                - phone
                - street
                - city
                - country
                - zipCode
            properties:
                businessName:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Nombre de la empresa
                    example: "Empresa ABC SAC"
                ruc:
                    type: string
                    minLength: 11
                    maxLength: 11
                    pattern: '^[0-9]{11}$'
                    description: Número RUC de la empresa
                    example: "20123456789"
                email:
                    type: string
                    format: email
                    description: Correo electrónico de la empresa
                    example: "contacto@empresaabc.com"
                phone:
                    type: string
                    minLength: 7
                    maxLength: 20
                    description: Número de teléfono
                    example: "+51 1 1234567"
                street:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Dirección de la calle
                    example: "Jr. Comercial 456"
                city:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Ciudad
                    example: "Lima"
                country:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: País
                    example: "Perú"
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 10
                    description: Código postal
                    example: "15002"
        UpdateEnterpriseRequest:
            type: object
            properties:
                businessName:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Nombre de la empresa
                    example: "Empresa ABC SAC Actualizada"
                email:
                    type: string
                    format: email
                    description: Correo electrónico de la empresa
                    example: "nuevo@empresaabc.com"
                phone:
                    type: string
                    minLength: 7
                    maxLength: 20
                    description: Número de teléfono
                    example: "+51 1 9876543"
                street:
                    type: string
                    minLength: 1
                    maxLength: 200
                    description: Dirección de la calle
                    example: "Jr. Nueva Dirección 789"
                city:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: Ciudad
                    example: "Trujillo"
                country:
                    type: string
                    minLength: 1
                    maxLength: 100
                    description: País
                    example: "Perú"
                zipCode:
                    type: string
                    minLength: 3
                    maxLength: 10
                    description: Código postal
                    example: "13001"
        CustomerResponse:
            type: object
            properties:
                id:
                    type: string
                    description: ID único del cliente
                    example: "987fcdeb-51a2-43d1-b789-123456789abc"
                name:
                    type: string
                    description: Nombre completo del cliente
                    example: "Empresa ABC SAC"
                customerType:
                    type: string
                    description: Tipo de cliente
                    example: "ENTERPRISE"
                documentNumber:
                    type: string
                    description: Número de documento
                    example: "20123456789"
                email:
                    type: string
                    format: email
                    description: Correo electrónico
                    example: "contacto@empresaabc.com"
                active:
                    type: boolean
                    description: Estado activo del cliente
                    example: true
                createdAt:
                    type: string
                    format: date-time
                    pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
                    description: Fecha de creación
                    example: "2024-01-15T10:00:00Z"
        ErrorResponse:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    description: Timestamp del error
                    example: "2024-01-15T10:00:00Z"
                status:
                    type: integer
                    description: Código de estado HTTP
                    example: 400
                error:
                    type: string
                    description: Tipo de error
                    example: "Bad Request"
                message:
                    type: string
                    description: Mensaje descriptivo del error
                    example: "Business name is required for enterprise customer"
                path:
                    type: string
                    description: Ruta donde ocurrió el error
                    example: "/enterprise"